+++ dev_scripts/lambda_handler.py
