+++ dev_scripts/tccc_analyzer.py
