+++ scripts/_1_4_8_estimated_evac_decision_times.py
@@ -1,4 +1,4 @@
-from misc_scripts.evac_response_window import compute_evac_windows
+from dev_scripts.evac_response_window import compute_evac_windows
 
 
 def main(mongo_db):
